+++ app/MyController.js
@@ -1,12 +1,13 @@
 exports.buildController = function (params) {
   let logPrefix = params.controllerName + '.controller';
   if (process.appParams.color) logPrefix = logPrefix.yellow;
-  return function (request, reqParams = {}, response) {
+  return async function (request, reqParams, response) {
+    reqParams = reqParams || {};
     request.logToConsole(logPrefix, reqParams);
 
     // make sure an admin is logged, or return an error page
-    reqParams.loggedUser = request.getUser();
-    if (params.adminOnly && !request.checkAdmin(response)) {
+    reqParams.loggedUser = await request.getUser();
+    if (params.adminOnly && !(await request.checkAdmin(response))) {
       console.log(logPrefix, 'must be logged as admin');
       return;
     }