+++ app/controllers/admin/FileController.js
@@ -1,10 +1,10 @@
 exports.buildController = function (params = {}) {
-  return function (request, reqParams, response) {
+  return async function (request, reqParams, response) {
     request.logToConsole(params.controllerName + '.controller', reqParams);
 
     // make sure an admin is logged, or return an error page
-    reqParams.loggedUser = request.getUser();
-    if (params.adminOnly && !request.checkAdmin(response)) return;
+    reqParams.loggedUser = await request.getUser();
+    if (params.adminOnly && !(await request.checkAdmin(response))) return;
 
     function render(res) {
       if (!res) response.badRequest();