+++ app/controllers/admin/testTemplates.js
@@ -7,11 +7,12 @@ const templateLoader = require('../../templates/templateLoader.js');
 
 const notifTemplates = require('../../templates/notif.js');
 
-exports.controller = function (request, reqParams = {}, response) {
+exports.controller = async function (request, reqParams, response) {
+  reqParams = reqParams || {};
   request.logToConsole('emailTemplate.controller', reqParams);
 
-  const user = request.checkAdmin(response);
-  if (false == user) return;
+  const user = await request.checkAdmin(response);
+  if (!user) return;
 
   const templateFile = reqParams.file;
   try {