+++ app/controllers/private/search.js
@@ -9,6 +9,7 @@ const mongodb = require('../../models/mongodb.js');
 const searchModel = require('../../models/search.js');
 const postModel = require('../../models/post.js');
 const followModel = require('../../models/follow.js');
+const userModel = require('../../models/user.js');
 const template = require('../../templates/search.js');
 
 const ObjectId = mongodb.ObjectId;
@@ -88,8 +89,8 @@ const fetchDataByType = {
     function next() {
       if (!idList.length) return cb(userList);
       const uid = '' + idList.shift();
-      postModel.fetchByAuthors([uid], { limit: 1 }, function (posts) {
-        const user = mongodb.usernames[uid];
+      postModel.fetchByAuthors([uid], { limit: 1 }, async function (posts) {
+        const user = await userModel.fetchAndProcessUserById(uid);
         if (user)
           userList.push({
             _id: uid,
@@ -265,7 +266,8 @@ function fetchSearchPage(myUid, q, cb) {
   });
 }
 
-exports.controller = function (request, reqParams = {}, response) {
+exports.controller = async function (request, reqParams, response) {
+  reqParams = reqParams || {};
   function renderSearchPage(q, cb) {
     fetchSearchPage(request.getUid(), q, function (results) {
       results.posts = results.tracks; // since algolia integration
@@ -359,7 +361,7 @@ exports.controller = function (request, reqParams = {}, response) {
   }
 
   request.logToConsole('search.controller', reqParams);
-  reqParams.loggedUser = request.getUser();
+  reqParams.loggedUser = await request.getUser();
 
   // track filter (from user profile)
   if (reqParams.q && reqParams.uid)