+++ cypress/support/e2e.ts
@@ -18,7 +18,6 @@
 /// <reference types="./" />
 
 import '@cypress/code-coverage/support';
-import '@applitools/eyes-cypress/commands';
 import './commands';
 
 beforeEach(function () {